fix(moe_int4_w4a16): cap default tile by N/K (MI300A LDS) + dtype-aware dequant (#85)
Two coupled root causes, both surfaced by the fp32 sweep shape {M:16,K:64,N:32}:

1. MI300A OutOfResources (Required 73728 > 65536 LDS). The production default
   config get_default_config returned [16,128,128,stages=2] UNCONDITIONALLY,
   bypassing prune_configs -- so for tiny N=32/K=64 it staged a tile bigger than
   the data: ~64 KiB LDS for bf16 (passes) but 72 KiB (73728) for fp32 (overflow,
   cdna3 LDS cap is 64 KiB/workgroup). Fix: get_default_config(M,N,K) now caps
   BLOCK_K at min(128,K) (power of two, multiple of pack-8 + group-32) and BLOCK_N
   at min(128, 2*N), mirroring prune_configs, and scales num_warps down for the
   smallest N tiles. For production shapes (K>=128) the caps are inert and the
   decode/prefill default [..,128,128] is unchanged -- no perf regression.

2. fp32 numerics drift (rel 9.06) on EVERY arch -- the issue attributed it to
   A100/GB10 separately, but it was masked on MI300A by the LDS crash. Root cause:
   the reference dequant_w4a16 rounded to bf16 UNCONDITIONALLY, so for fp32
   activations the reference used bf16-rounded weights while the Triton kernel
   dequants to a.dtype (fp32). The kernel was MORE precise than the oracle by
   ~bf16-eps (~4e-3 abs, rel ~9 at near-zero outputs). Fix (mixed-precision
   principle: reference is the high-precision source of truth): dequant_w4a16 now
   takes out_dtype; the reference passes A.dtype so dequant precision MATCHES the
   kernel -- fp32 activations keep the full-precision product (no lossy bf16
   round), bf16 activations round to bf16 (unchanged). The kernel needed no change
   (it already cast to a.dtype). The unit test's _ref_grouped oracle is updated
   to the same dtype-aware dequant (it was passing before only by error
   cancellation).

Verified (verify + verify_parity, every sweep shape):
  MI300A (amd_cdna3):  compiled+passed (failing=0), parity agree
  GB10   (nvidia_sm121): compiled+passed (failing=0), parity agree
  A100   (nvidia_sm80):  compiled+passed (failing=0), parity agree
CPU interpreter: 104/104 moe tests pass (the 1 failure is a CUDA-graph-capture
test, GPU-only, unrelated). Production perf unchanged (caps inert for K>=128).

Closes #85 (the LDS retile AND the numerics line the issue deferred -- same root
cause, masked on MI300A by the crash).

+++ src/xkernels/ops/moe/reference.py
@@ -52,7 +52,7 @@ def moe_w4a16_ref(
     Returns:
         ``[M, N]`` output in ``A.dtype`` (fp32 accumulation).
     """
-    W = dequant_w4a16(packed, scale, group_size)  # [E_local, N, K] bf16
+    W = dequant_w4a16(packed, scale, group_size, out_dtype=A.dtype)  # [E_local, N, K]
     M, topk = topk_ids.shape
     out = torch.zeros(M, W.shape[1], dtype=torch.float32, device=A.device)
     emap = None if expert_map is None else expert_map.to(A.device)
